Move AnimationComponent to ecs.components.animation
