+++ game/src/ecs/items/ItemData.java
@@ -1,11 +1,11 @@
 package ecs.items;
 
 import configuration.ItemConfig;
-import ecs.components.AnimationComponent;
 import ecs.components.HitboxComponent;
 import ecs.components.InventoryComponent;
 import ecs.components.ItemComponent;
 import ecs.components.PositionComponent;
+import ecs.components.animation.AnimationComponent;
 import ecs.components.stats.DamageModifier;
 import ecs.entities.Entity;
 import graphic.Animation;
