+++ game/src/ecs/systems/DrawSystem.java
@@ -1,8 +1,8 @@
 package ecs.systems;
 
-import ecs.components.AnimationComponent;
 import ecs.components.MissingComponentException;
 import ecs.components.PositionComponent;
+import ecs.components.animation.AnimationComponent;
 import ecs.entities.Entity;
 import graphic.Animation;
 import graphic.Painter;
