+++ game/test/ecs/systems/DrawSystemTest.java
@@ -3,9 +3,9 @@ package ecs.systems;
 import static org.junit.Assert.assertThrows;
 
 import controller.SystemController;
-import ecs.components.AnimationComponent;
 import ecs.components.MissingComponentException;
 import ecs.components.PositionComponent;
+import ecs.components.animation.AnimationComponent;
 import ecs.entities.Entity;
 import graphic.Animation;
 import graphic.Painter;
